moved truck speed test, renamed truck restrictions test
